Added ImageUploadeHelper in utils
